[DEV] Ground SDK Tools 8.5 release

+++ src/mbuf_audio_frame.c
@@ -381,7 +381,7 @@ int mbuf_audio_frame_copy(struct mbuf_audio_frame *frame,
 
 	ssize_t tmp = mbuf_audio_frame_get_size(frame);
 	if (tmp < 0)
-		return tmp;
+		return (int)tmp;
 	size_t required_len = tmp;
 
 	ret = mbuf_base_frame_rdlock(&frame->base);
